requirements.txt
click==8.1.7
colorama==0.4.6
markdown-it-py==3.0.0
mdurl==0.1.2
pillow==10.4.0
pillow-avif-plugin==1.4.6
Pygments==2.18.0
pywin32==306
rich==13.7.1
shellingham==1.5.4
typer==0.12.3
typing_extensions==4.12.2
